// src/types/user-dataType.ts
export type UserDataType = {
  user_id: string;
  avatar: string;
  first_name: string;
  last_name: string;
  email: string;
  phone_number: string;
  country: string;
  state: string;
  city: string;
  pincode: string;
  company_name: string;
  branch_name: string;
  globle_status_name: string;
  role_name: string;
is_active?: boolean;
license_no?:string;
}